+++ react_dars_website/src/App.js
@@ -1,15 +1,23 @@
-/* 
-import logo from './logo.svg';
 import './App.css';
-import { useEffect, useState } from 'react';
-*/
-import LoginForm from './Components/LoginForm/loginForm.js';
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import About from "./Components/Pages/About.js";
+import LoginForm from './Components/Pages/login.js';
+import Temp from './Components/Pages/Temp.js';
 
 function App() {
 
 
   return (
-      <LoginForm />
+    <div className = "App">
+      <Router>
+          <Routes>
+            {/* rount for each page */}
+            <Route path="/" element={<LoginForm />} />
+            <Route path="/about" element={<About />} />
+            <Route path="/temp" element={<Temp />} />
+          </Routes>
+      </Router>
+    </div>
   );
 }
