+++ react_dars_website/src/Components/Pages/Temp.js
@@ -1,0 +1,20 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+
+function Temp(){
+    const navigate = useNavigate();
+
+    return (
+        <div className="visualBG">
+            <h1>Temp Page</h1>
+            <button 
+                onClick={() => {
+                    navigate("/about");
+                }}> 
+                Go to the About page 
+            </button>
+        </div>
+    );
+}
+
+export default Temp;
