+++ react_dars_website/src/index.css
@@ -15,7 +15,7 @@ body {
   justify-content: center;
   align-items: center;
   min-height: 100vh;
-  background: url('./Components/Assets/loginWP.jpg') 
+  background: /*url('./Components/Assets/loginWP.jpg')*/  white
   no-repeat;
   background-size: cover;
   background-position: center;
